src/requirements.txt
fastapi==0.103.0

redis~=5.0.0
PyJWT~=2.8.0